Fixed warning: comparison between signed and unsigned integer expressions
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@1981 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/src/protomol/force/hessian/ReducedHessGB.cpp
@@ -267,7 +267,7 @@ Real ReducedHessGB::FirstDerivativeFGB(
    
   unsigned int sz = topo->atoms.size();
 
-  for(int k=0; k<sz; k++){
+  for(unsigned k=0; k<sz; k++){
     
     if( k == atom1 ) continue;
       
@@ -333,9 +333,9 @@ ReducedHessGB::secondDerivativeRawTerms ReducedHessGB::SecondDerivativeFGB(
 
   unsigned int sz = topo->atoms.size();
   
-  for(unsigned int k=0; k<sz; k++){
+  for(unsigned k=0; k<sz; k++){
     
-    if( k == atom1 ) continue; //|| k == atom2 
+    if( k == (unsigned)atom1 ) continue; //|| k == atom2 
 
     //grab r_{ik} / r_{jk}
     const Real dist = topo->atoms[atom1].myGBSA_T->distij[k];
